Added GetValueAsString method with a parameter for a client-controlled buffer instead of the static temporary.
The static temporary has unclear owner and lifespan semantics and should be deprecated. It's hard to change due to the const method declaration (though the const isn't the problem, it's the local static). Removing the const has backward compatibility implications.

The static temp also puts a heap allocation's lifespan out of reach of the client.

Most of the time this isn't an issue -- why doesn't the client just use get value and convert to the string by itself? This works when working with the concrete types (int, float, identifier, string) but not with the base wme element class. Certainly possible to do the poor-man's rtti the WMElement interface provides but this method already exists on that same interface so why not make it usable with well-defined memory semantics.

git-svn-id: https://soar.googlecode.com/svn/trunk@13000 915bf4ee-6d5f-11de-9cf4-4bf9f06b8206