fixes the bug introduced in r12997
The previous versions worked because the ref counts were integers, not sets. When the ref count for an id is checked, the count defaults to 0 if the key (ie. the id) is not in the map, which is exactly the desired behavior. For sets, however, this becomes a NULL pointer, which causes a segfault. The fix is to make sure that the refcount for that id exists first.



git-svn-id: https://soar.googlecode.com/svn/trunk@13091 915bf4ee-6d5f-11de-9cf4-4bf9f06b8206