Fixing the merge in of the new schema branch.  Here are the full commits from that branch ----------------------------------------- Init/Reinit Changes - Fixed epmem bug 125 where epmem recording breaks after init-soar.  Soar will now properly close and re-initialize epmem. - Soar will now also close smem on init-soar - Added epmem --init command.  This will close and re-initialize the database like smem --init.  Made sure everything is cleaned up properly. - Removed epmem -close command.  No longer needed with above two changes.  Close function exists internally and used by other code. - Removed restrictions on changing smem/epmem database paths, database storage method (memory vs disk) and append mode.  We can now change those setting at any time.  When epmem/smem is re-initialized or an init-soar is issued, Soar will use the new settings. - Soar will now no longer automatically switch to database mode the first time that you set the path.
UI Changes
- Added -on, -off, -e, -d, --enable, --disable parameters to epmem and smem to make interface consistent with other systems.
- Fixed cli bug where parameter checking in cli_Commands.h would not print error information if parameters were invalid.  (It was building an error string, just not printing it.)
- Fixed smem bug where Soar would not tell user that certain smem -add commands were bad (for example smem -add {fdsfds})
- Converted all epmem/smem/learning menus and output to use new consistent UI style
- Added source file feedback.  Fixed linefeed after final "*" or "#".
- Added a bunch of feedback messages to epmem, smem and learn commands.  Made the existing feedback consistent.
- Changed more of smem to use new trace command instead of random ways
- Updated cli help with new command changes and recompiled into kernel
- Changed CommandLineInterface::PrintCLIMessage_Item function to take any soar_module::named_object so that it can be used to print stats and timers

Code Cleanup
- Fixed issues with parts of source code putting linefeeds at the beginning of trace messages instead of the end like most parts of the code.  (I'm sure there's are more.)
- Added DEBUG_USE_STDERR_TRACE so that trace print function will send to stderr instead of  the print output handler or xml, which can have issues when debugging
- Added several PrintCLIMessage functions so that the various CLI commands can produce consistent output.  So far,  epmem, smem, and learn have been converted to use the functions.  Also handles the logic to send output via SML so it makes the code for the individual cli command classes much cleaner.
- Added some comments to rete.cpp

Schema Changes
- Changed some more comments in epmem code that used the node/edge terminology to be more consistent with new schema.
- Ran regression and performance tests on smem schema changes.

Other Epmem Changes
- Changed default epmem recording trigger to be decision cycle
- Filled epmem feature request 123: ability to append database

-- Older Commits -- 

Fixed bug where CLI would exit without shutting down the kernel.  This could
result in data not being written to the episodic and/or semantic memory
databases.
---------------------
- Smem refactored for new schema.
- Added append-database options to smem and epmem.  If the setting is off and
the database is on disk, Soar will dump the data in the database upon
initialization. Otherwise, it will append new data to database.  Epmem has
append off by default while smem has it on.
- Added robust version checking for epmem and smem schemas.  Both system use
identical version checking approach in case we merge the two databases.
- If there any version check failures, Soar will now print a message and
automatically switch to a memory-based database instead.  Cleaner than previous
solution.
- Smem statement and table initialization changed somewhat to support db append
setting and switching to memory.
- Added some sql utility functions common to epmem and smem to soar_db.  There's
probably more we can move later.
- Added a trace printing function to print.cpp and changed smem and epmem trace
messages to use them
---------------------
- Refactored code and new code for new schema fully working now.
- Print and visualize episodes fixed.
- Removed last remnants of temporal hash table
---------------------
Epmem will now lookup constants from the four symbol lookup tables instead of
the old temporal symbol hash table.  First working version.  Print and visualize
episode code still hasn't been updated though yet.
---------------------
How symbol type is handled is changed somewhat  from how smem does it and epmem
previously did it;  that code retrieves the symbol types within the original
large query. The theory behind this change is that it we'll gain more from
removing a join from the big, more computationally intensive query than we lose
from the overhead of a second query.  This leverages the fact that id's and
attributes are always a known type, so we don't even need to join with the type
table for those. (and reverse hashing for attributes is probably more common
than for values with the way the code/tables are structured.)  We'll see if
there's any truth to this when testing.
---------------------
- Added split up symbol hash tables to epmem (creation, init, statements)
- Added code so epmem stores symbol lookup data in 4 tables (in addition to old
single temporal hash table)
- Converted smem hash utility functions for epmem
- Moved symbol_is_constant to symtab.h instead of semantic_memory.cpp so that
epmem can use it too
---------------------
- Refactored epmem database for basics of new schema. (does not split up
temporal hash table yet)
- Instrumented _epmem_store_level heavily to better understand what's going on
with reservations, the various id's that epmem uses and how it adds wme's to the
wmg. Enabled with #DEBUG_EPMEM_ADD_WMES but will probably be removed later
- Added new node table and add statement.  Added code so that when Soar starts
using a new n_id (node id but not to be confused with previous use of node as a
wme with a constant as its value), it adds that id to the node table.  This
table is not used anywhere, but makes the database relations strict.  The root
node is now no longer a special case.
- Added functions to trace and profile sql statements and registered callbacks
with sqlite code.  Enabled with #DEBUG_EPMEM_SQL


git-svn-id: https://soar.googlecode.com/svn/trunk@13788 915bf4ee-6d5f-11de-9cf4-4bf9f06b8206